move import from beerxml to lib. make attenuation and cultureDate optional.
