+++ src/types/yeast.js
@@ -2,10 +2,10 @@
 export type Yeast = {
   name: string,
   amount: number,
-  attenuation: number,
+  attenuation?: number,
   form: 'Liquid' | 'Dry' | 'Slant' | 'Culture',
   type: 'Ale' | 'Lager' | 'Wheat' | 'Wine' | 'Champagne',
-  cultureDate: string
+  cultureDate?: string
 }
 
 export const YeastTypes = {
